% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{ft_schema_flat}
\alias{ft_schema_flat}
\title{Print a flat filetree schema summary}
\usage{
ft_schema_flat(ft)
}
\arguments{
\item{ft}{A \code{filetree} object.}
}
\value{
The input \code{filetree}, invisibly.
}
\description{
\ifelse{html}{\out{<span class="badge badge-warning">Experimental</span>}}{\strong{Experimental}}
}
\details{
Print the result of \code{\link[=ft_format_schema_flat]{ft_format_schema_flat()}} and invisibly return the input
\code{filetree}.
}
\examples{
root <- system.file("demo-1", package = "filetree")
ft <- ft_init(root, c("subject", "time", "data"))

ft_schema_flat(ft)
}
