% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{ft_schema_tree}
\alias{ft_schema_tree}
\title{Print a filetree schema tree}
\usage{
ft_schema_tree(ft)
}
\arguments{
\item{ft}{A \code{filetree} object.}
}
\value{
The input \code{filetree}, invisibly.
}
\description{
Print the result of \code{\link[=ft_format_schema_tree]{ft_format_schema_tree()}} and invisibly return the input
\code{filetree}.
}
\examples{
root <- system.file("demo-1", package = "filetree")
ft <- ft_init(root, c("subject", "time", "data"))

ft_schema_tree(ft)
}
