% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{ft_set_root}
\alias{ft_set_root}
\title{Change a filetree root}
\usage{
ft_set_root(ft, root)
}
\arguments{
\item{ft}{A \code{filetree} object.}

\item{root}{Path to the new root directory used as the base for indexing.}
}
\value{
The updated \code{filetree} object.
}
\description{
Return a copy of a \code{filetree} object with a new root directory.
}
\examples{
root <- system.file("demo-1", package = "filetree")
ft <- ft_init(root, c("subject", "time", "data"))

ft_set_root(ft, tempdir())
}
