% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{print.filetree}
\alias{print.filetree}
\title{Print a filetree summary}
\usage{
\method{print}{filetree}(x, ..., width = getOption("width"))
}
\arguments{
\item{x}{A \code{filetree} object.}

\item{...}{Unused, included for method signature compatibility.}

\item{width}{Optional output width forwarded to \code{\link[=format.filetree]{format.filetree()}}.}
}
\value{
The input \code{filetree} object, invisibly.
}
\description{
Print the formatted summary of a \code{filetree} object to the console.
}
\examples{
root <- system.file("demo-1", package = "filetree")

ft <- ft_init(root, c("subject", "time", "data"))

print(ft)
}
