% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree-package.R
\docType{package}
\name{filetree-package}
\alias{filetree}
\alias{filetree-package}
\title{filetree: Declarative File Tree Schemas and Validation}
\description{
Provides a 'filetree' object for declaring the expected structure of a file hierarchy. A file listing can be parsed and validated against the declared file tree to extract metadata and identify structural and naming inconsistencies.
}
\examples{
root <- system.file("demo-1", package = "filetree")
ft <- ft_init(root, c("subject", "time", "data"))
ft
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/filetree}
  \item Report bugs at \url{https://github.com/tjmahr/filetree/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
