% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree-package.R
\docType{package}
\name{filetree-package}
\alias{filetree}
\alias{filetree-package}
\title{filetree: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
