% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{ft_add_regex}
\alias{ft_add_regex}
\title{Register field regexes used by component templates}
\usage{
ft_add_regex(ft, regexes)
}
\arguments{
\item{ft}{A \code{filetree} object.}

\item{regexes}{Named character vector of regular expressions to store.}
}
\value{
The updated \code{filetree} object.
}
\description{
Add named regular expressions for values that should be extracted from
directory or file names. Field regexes can be referenced by placeholders such
as \code{{subject}} inside directory or file component templates. Field regexes can
also reference other field regexes with the same placeholder syntax.
}
\examples{
root <- system.file("demo-1", package = "filetree")

ft <- ft_init(root, c("subject", "time", "data")) |>
  ft_add_regex(c(
    wtocs = "wT\\\\d\\\\d",
    stocs = "s\\\\dT[01]\\\\d",
    tocs = "{wtocs}|{stocs}",
    subject = "\\\\w{2}-\\\\d{2}",
    time = "day\\\\d{2}",
    task = "red|green"
  ))

names(ft$regex_pool)
}
