% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{ft_add_regex}
\alias{ft_add_regex}
\title{Register regex templates used by patterns}
\usage{
ft_add_regex(ft, regexes)
}
\arguments{
\item{ft}{A \code{filetree} object.}

\item{regexes}{Named character vector of regular expressions to store.}
}
\value{
The updated \code{filetree} object.
}
\description{
Add named regular expressions to the pool that can be referenced by
placeholders such as \code{{subject}} inside directory or file patterns.
}
