% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetree.R
\name{ft_list}
\alias{ft_list}
\title{List files under the filetree root}
\usage{
ft_list(ft, include_ignored = FALSE)
}
\arguments{
\item{ft}{A \code{filetree} object.}

\item{include_ignored}{Logical. If \code{TRUE}, return files that match ignored
file or directory templates. If \code{FALSE}, ignored files are pruned.}
}
\value{
An \code{fs_path} character vector of file paths under \code{ft$root}.
}
\description{
Return all files under the filetree root using the configured \code{fs} helper.
}
\examples{
root <- system.file("demo-1", package = "filetree")

ft <- ft_init(root, c("subject", "time", "data"))

head(ft_list(ft))
}
